requirements.txt
asgiref==3.2.3
Django==3.0.1
django-ckeditor==5.8.0
django-cors-headers==3.2.0
django-filter==2.2.0
django-js-asset==1.2.2
django-material-admin==1.6.23
djangorestframework==3.11.0
Markdown==3.1.1
Pillow==6.2.1
pytz==2019.3
sqlparse==0.3.0
